requirements.txt
opencv-python>=4.9.0.80
numpy>=1.26.0
Pillow>=10.0.0
torch>=2.0.0
transformers>=4.30.0
face_recognition
dlib>=4.30.0
# Add to requirements.txt
chromadb>=0.4.0
sentence-transformers>=2.2.2
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.5
jinja2>=3.1.2
python-telegram-bot>=20.7
python-dotenv>=1.0.0